requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
joblib
numpy
scikit-learn==1.6.1